requirements.txt
python-telegram-bot==13.15
instaloader==4.10.1
requests==2.28.1
python-dotenv==0.21.1
pytz